fix(snell-rollcall): reach every node of a real controller, ask for what it has
Closes #1026.

Found by running the vendor Centra controller — the simulator already in
assets/, as a Sirius 800 — and pointing our own consumer at it. It could not
open a session at all, and once it could, fourteen of the controller's fifteen
nodes were unreachable. Both were our defects. The measurements are in
internal/snell-rollcall/docs/oracle-centra.md.

Services. We asked every peer for a fixed set including SV_DISPLAY. This
controller has no display service and refuses the whole call, because services
are all-or-nothing; the identical call without it succeeds immediately before
and after, so it is the service and not the controller's mood. The mask is now
what we want intersected with what the peer advertised, taken per node: each
entry in a device list carries its own services and they differ between the
matrices, the panel node and the gateway.

Addressing. A slot was a port on the gateway's own unit. On a controller the
nodes differ by unit — Matrix 1 at 0000-11-00, the XY Panel that serves the
routing interface at 0000-81-00 — so a slot read went to the wrong place and
timed out. A slot is now a position in the device's own enumeration and its
address is whatever the device put there: a port on a frame, a unit on a
controller. The address is reported in the slot's identity, because on a
controller a slot number is a position in a list rather than a place in a
frame, and an operator needs to see which node it reached.

The device map has its own session. A request belonging to a service the
session did not negotiate is answered SP_INVSESS, not ignored, so enumerating
on the control session failed. The map service is asked for alone and without
the long-string bit, which this controller refuses in combination and which
buys nothing on a fixed-width structure; a gateway that will not open one is
asked on the control session instead, since some units answer a list on any
session.

The same run validated codec/router against the controller: every offset in
every sub-table, the four table sizes, the source-pin packing, and the
names-file checksum, which reproduces the controller's published values
exactly. The package comment no longer calls them unvalidated. A live route was
set and read back, which is how the tally-follows-tielines behaviour and the
set reply carrying the previous value were established.

Coverage stays at 100% of statements across all six packages.

+++ internal/snell-rollcall/consumer/value.go
@@ -161,7 +161,7 @@ func (p *Plugin) SetDefault(ctx context.Context, req consumer.ValueRequest) (con
 // landing in the gap between the two calls, which is a branch nothing can
 // exercise and nobody can reason about.
 func (p *Plugin) locate(ctx context.Context, req consumer.ValueRequest) (*menuLine, *session.Session, error) {
-	s, err := p.session(ctx, uint8(req.Slot))
+	s, err := p.session(ctx, req.Slot)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -274,7 +274,7 @@ func (p *Plugin) encodeValue(line *menuLine, val consumer.Value) (codec.Mode, in
 // and a warning. That is why they have their own accessor rather than
 // appearing in a walk.
 func (p *Plugin) Display(ctx context.Context, slot int) (map[int16]string, error) {
-	s, err := p.session(ctx, uint8(slot))
+	s, err := p.session(ctx, slot)
 	if err != nil {
 		return nil, err
 	}
